Update to model_configs for hyperparam grid for trraining and ORM structure, updated model_evaluation for centrally managing metrics, dependency clean-up
